feat: revoke and certificate services

+++ backend/src/main/java/com/example/demo/repository/CertificateRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.model.CertificateData;
 
 @Repository
-public interface CertificateRepository extends JpaRepository<CertificateData, Integer>{
+public interface CertificateRepository extends JpaRepository<CertificateData, Long>{
 	
 	@Query("SELECT cd FROM CertificateData cd WHERE cd.serialCode=?1")
 	public CertificateData getCertificateByCode(String serialCode);
